server/requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
bidict==0.23.1
click==8.1.8
exceptiongroup==1.3.1
fastapi==0.128.8
h11==0.16.0
idna==3.11
pydantic==2.12.5
pydantic_core==2.41.5
python-engineio==4.13.1
python-socketio==5.16.1
simple-websocket==1.1.0
starlette==0.49.3
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.39.0
wsproto==1.2.0
